Modifiche al middleware di autenticazione

+++ Deliverables/Implementation/quiad_rest_api_server/src/tree/services/tree.service.ts
@@ -7,9 +7,14 @@ export class TreeService {
     private nodeController = new NodeController();
 
     public async getNodes(req: Request, res: Response, next: NextFunction) {
+        const account = res.locals.account;
         const owner = parseInt(req.params.owner);
-        const nodes = await this.nodeController.getNodes(owner);
-        res.json(nodes);
+        if(account.id == owner) {
+            const nodes = await this.nodeController.getNodes(owner);
+            res.json(nodes);
+        } else {
+            res.status(401).json(null);
+        }
     }
 
     public async createNode(req: Request, res: Response, next: NextFunction): Promise<void> {
@@ -31,31 +36,41 @@ export class TreeService {
     }
 
     public async deleteNode(req: Request, res: Response, next: NextFunction): Promise<void> {
+        const account = res.locals.account;
+        const node = account.user.node;
         const id = parseInt(req.params.id);
-        try {
-            const deletedNode = await this.nodeController.deleteNode(id);
-            if(deletedNode == null) {
-                res.status(404).json(null);
-            } else {
-                res.json(deletedNode);
-            }
-        } catch(err) {
+        if(node.id == id) {
             res.status(500).json(null);
+        } else {
+            try {
+                const deletedNode = await this.nodeController.deleteNode(id);
+                res.json(deletedNode);
+            } catch(err) {
+                res.status(500).json(null);
+            }
         }
     }
 
     public async bindDocument(req: Request, res: Response, next: NextFunction): Promise<void> {
         const node =  parseInt(req.params.node);
         const document = parseInt(req.params.document);
-        const bindedNode = await this.nodeController.bindDocument(node, document);
-        res.json(bindedNode);
+        try {
+            const bindedNode = await this.nodeController.bindDocument(node, document);
+            res.json(bindedNode);
+        } catch(err) {
+            res.status(500).json(null);
+        }
     }
 
     public async unbindDocument(req: Request, res: Response, next: NextFunction): Promise<void> {
         const node =  parseInt(req.params.node);
         const document = parseInt(req.params.document);
-        const unbindedNode = await this.nodeController.unbindDocument(node, document);
-        res.json(unbindedNode);
+        try {
+            const unbindedNode = await this.nodeController.unbindDocument(node, document);
+            res.json(unbindedNode);
+        } catch(err) {
+            res.status(500).json(null);
+        }
     }
 
 }
